refactor: move transaction stream
