+++ ethers-providers/src/provider.rs
@@ -22,12 +22,6 @@ use serde::{de::DeserializeOwned, Serialize};
 use thiserror::Error;
 use url::{ParseError, Url};
 
-use futures_core::{Future, Stream};
-use futures_util::stream::FuturesUnordered;
-use futures_util::{StreamExt, TryFutureExt};
-use std::collections::VecDeque;
-use std::pin::Pin;
-use std::task::{Context, Poll};
 use std::{convert::TryFrom, fmt::Debug, time::Duration};
 use tracing::trace;
 use tracing_futures::Instrument;
@@ -988,73 +982,5 @@ mod tests {
             .collect::<Vec<_>>()
             .await;
         assert_eq!(blocks, vec![1, 2, 3]);
-    }
-}
-
-type TransactionFut<'a> = Pin<Box<dyn Future<Output = TransactionResult> + 'a>>;
-
-type TransactionResult = Result<Transaction, ProviderError>;
-
-struct TxStream<'a, P> {
-    pending: FuturesUnordered<TransactionFut<'a>>,
-    buffered: VecDeque<TxHash>,
-    provider: &'a Provider<P>,
-    watcher: FilterWatcher<'a, P, TxHash>,
-    max_concurrent: usize,
-}
-
-impl<'a, P: JsonRpcClient> Stream for TxStream<'a, P> {
-    type Item = TransactionResult;
-
-    fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let this = self.get_mut();
-
-        if let tx @ Poll::Ready(Some(_)) = this.pending.poll_next_unpin(cx) {
-            return tx;
-        }
-
-        while this.max_concurrent < this.max_concurrent {
-            if let Some(_) = this.buffered.pop_front() {
-                // TODO add get_tx again
-            } else {
-                break;
-            }
-        }
-
-        let mut watcher_done = false;
-        loop {
-            match Stream::poll_next(Pin::new(&mut this.watcher), cx) {
-                Poll::Ready(Some(tx)) => {
-                    if this.pending.len() < this.max_concurrent {
-                        this.pending
-                            .push(Box::pin(this.provider.get_transaction(tx).and_then(
-                                |res| {
-                                    if let Some(tx) = res {
-                                        futures_util::future::ok(tx)
-                                    } else {
-                                        futures_util::future::err(ProviderError::CustomError(
-                                            "Not found".to_string(),
-                                        ))
-                                    }
-                                },
-                            )));
-                    } else {
-                        this.buffered.push_back(tx);
-                    }
-                }
-                Poll::Ready(None) => {
-                    watcher_done = true;
-                    break;
-                }
-                _ => break,
-            }
-        }
-
-        if watcher_done && this.pending.is_empty() {
-            // all done
-            return Poll::Ready(None);
-        }
-
-        Poll::Pending
     }
 }
